refactor: game 패키지 추가 및 Color의 changeColor에서 Blank 처리
Color changeColor에 Blank가 전달되는 경우 Blank를 반환
ALL 패키지 변경에 따른 리팩토링
