+++ src/main/java/chess/domain/piece/movable/KingMovable.java
@@ -1,6 +1,5 @@
 package chess.domain.piece.movable;
 
-import chess.domain.*;
 import chess.domain.board.Board;
 import chess.domain.board.position.Column;
 import chess.domain.board.position.Position;
